move instanceType
