+++ models/common/instance.ts
@@ -1,4 +1,4 @@
-import type { IInstanceType } from '~/models/instanceType';
+import type { IInstanceType } from '~/models/instances/instanceType';
 
 export type IInstance = {
   type: IInstanceType;
