+++ repositories/repositoryFactory.ts
@@ -1,4 +1,4 @@
-import type { IInstanceType } from '~/models/instanceType';
+import type { IInstanceType } from '~/models/instances/instanceType';
 import { mastodonRepository } from '~/repositories/instances/mastodon';
 import { misskeyRepository } from '~/repositories/instances/misskey';
